R6 class storing health statistics for a list of health units.

R6 class storing health statistics for a list of health units.

R6 class storing health statistics for a list of health units.

Public fields

.data

A list of health units in health unit grp.

.sf

A sf storing spatial information.

.id

A character ID of object.

Methods

Public methods

health_unitgrp$new()

health_unitgrp$validate()

health_unitgrp$metadata()

health_unitgrp$data()

health_unitgrp$combine_data()

health_unitgrp$sf()

health_unitgrp$ID()

health_unitgrp$IDs()

health_unitgrp$titles()

health_unitgrp$health_unit()

health_unitgrp$subset()

health_unitgrp$get_download()

health_unitgrp$download_handler()

health_unitgrp$clone()

Method new()

Create instance of health unit grp.

Usage

health_unitgrp$new(.data, .sf, .id)

Arguments

.data

(list)
A list of health units in health unit grp.

.sf

(sf) A sf storing spatial information for health units.

.id

(characer) A character ID of object.

Method validate()

Validate structure of health unit grp.

Usage

health_unitgrp$validate()

Method metadata()

Get combine metadata of health unit grp.

Usage

health_unitgrp$metadata()

Method data()

Get data of health unit grp.

Usage

health_unitgrp$data()

Method combine_data()

Get combined metadata and data in single data.frame.

Usage

health_unitgrp$combine_data()

Method sf()

Get sf of health unit grp.

Usage

health_unitgrp$sf()

Method ID()

Get ID of health unit grp.

Usage

health_unitgrp$ID()

Method IDs()

Get IDs of stored health units

Usage

health_unitgrp$IDs()

Method titles()

Get names of stored health units

Usage

health_unitgrp$titles()

Method health_unit()

Get stored health unit.

Usage

health_unitgrp$health_unit(id)

Arguments

id

(character(1))
Character specifying ID of object to obtain from group.

Method subset()

Subset health unit group.

Usage

health_unitgrp$subset(id)

Arguments

id

(character())
Character specifying ID (or IDs) of object to obtain from group.

Method get_download()

Get downloadable data.frame of health unit group.

Usage

health_unitgrp$get_download()

Method download_handler()

Get download handler function of health unit group.

Usage

health_unitgrp$download_handler()

Method clone()

The objects of this class are cloneable with this method.

Usage

health_unitgrp$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.